requirements.txt
fastapi
pymongo
pydantic
snowflake.connector
pandas
marshmallow
dash
fbprophet
sklearn